Checking in, getting interesting
What's fun at this point is that we're hitting a design challenge.
The purpose facet's onlyOwner modifier has revealed that the owner is not set.
So now we have to handle how do owned facets work?
Does each owned facet get initialized with an owner? That would seem unlikely.
More likely I need to add an Ownership facet, and compose that with the purpose facet.
This is becoming a more and more rich example, which is cool b/c it means producing more useful code.